01APIForge/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Redis
redis==5.0.1
aioredis==2.0.1

# ZeroMQ
zmq==0.0.0
pyzmq>=26.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt>=4.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
loguru==0.7.2
email-validator==2.0.0.post2

# Cloud SDKs
supabase==2.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# API Documentation
fastapi-users==12.1.2
fastapi-users[sqlalchemy]==12.1.2
